services/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
authors = [
    {name = "Jakub Kropáček", email = "kropikuba@gmail.com"},
    {name = "Nikola Kubeczková", email = "kubeczkova.n@gmail.com"}
]
dependencies = [
    "django>=5.1.5",
    "psycopg2-binary>=2.9.10",
    "uvicorn[standard]>=0.34.0",
    "djangorestframework>=3.15.2",
    "django-debug-toolbar>=5.0.1",
    "django-environ>=0.12.0",
    "wait-for-it>=2.3.0",
    "django-cors-headers>=4.7.0",
    "django-post-office>=3.9.1",
    "pytest-django>=4.10.0",
    "factory-boy>=3.3.3",
    "pytest-cov>=6.0.0",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
